Refactor Fragment classes into their own directory
