+++ app/src/main/java/com/example/fitnesslog/history/view/fragment/HistoryFragment.kt
@@ -1,10 +1,10 @@
-package com.example.fitnesslog.history.view
+package com.example.fitnesslog.history.view.fragment
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.example.fitnesslog.R
 
 // TODO: Rename parameter arguments, choose names that match
