+++ app/src/main/java/com/example/fitnesslog/program/view/fragment/ProgramFragment.kt
@@ -1,10 +1,10 @@
-package com.example.fitnesslog.workout.view
+package com.example.fitnesslog.program.view.fragment
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.example.fitnesslog.R
 
 // TODO: Rename parameter arguments, choose names that match
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [WorkoutFragment.newInstance] factory method to
+ * Use the [ProgramFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class WorkoutFragment : Fragment() {
+class ProgramFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class WorkoutFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_workout, container, false)
+        return inflater.inflate(R.layout.fragment_program, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class WorkoutFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment WorkoutFragment.
+         * @return A new instance of fragment ProgramFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            WorkoutFragment().apply {
+            ProgramFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
