+++ app/src/main/res/layout/fragment_history.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".history.view.HistoryFragment">
+    tools:context=".history.view.fragment.HistoryFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
